Adding lowgc quadtree
Here each of the parent->child relations uses an integer based lookup
instead of a direct pointer.

The tree itself still contains pointers, as the data is held in a []T
slice. The next step is to eliminate this pointer to allow the tree to
have almost no GC impact.