Added tests for the byte/object stores
We also changed the interface of the ByteStore. Previously creating a
new byte slice took a byte slice as an argument. The contents of this
slice would be copied into the new byte slice inside the store. But this
means that the user must _always_ create a byte slice in order to copy
those values into the stored slice.

It's likely that there are many scenarios where an initial byte slice is not
needed and requiring one would force unecessary allocations and copying.

Now the ByteStore.New(...) only asks for the size of the slice to be
created.