Parsing/reading errs are inlined into RawCSVEntry
This means that errors inside the async flow can be read by the end
consumer.

By having LineNum and Error directly in the RawCSVEntry struct it does
feel like we are mixing concerns here. It might be cleaner if we
separate these two fields into a separate struct and then bind them
together in a higher level struct. This will only change reabibility.