Adding validation to long/lat values
We now check that the long/lat values parsed in lds_csv are valid. This
was added after we observed longitude values of ~183 which is outside
the -180...180 range. It's likely that these outside values can be
normalised by wrapping them around to -180. For now we just consider
these values to be errors.