Added ByteStore.Free()
We have used the same meta-data based approach that we used in
ObjectStore. However there was a complication, in that the offset which
points into a slice of bytes can't be used to point into a slice of
meta-data objects. So we now maintain two offsets, byteOffsets and
slotOffsets.

One important thing to note, is that we removed the writing of the size
of an allocation from the byte slices themselves. This piece of
meta-data was unncessary because we were storing the size of a slice in
the pointer. This significantly simplifies the implementation. One thing
that's worth pondering here is that this was missed initially. There was
never any reason to write the size into the slice itself, but this (in
hindsight) observation was missed. This highlights the importance of
full system comprehension in programming, if you start dropping details
out of your brain, you are much more likely to produce over-complicated
implementations, which are then even harder to keep inside your brain.
Although this is a small error, and easily fixed, it feels like a
warning. A likely worthwhile immediate step is a series of careful
renamings, to ensure that at least this part is clear.