Simplify small sized byte slabs
Previously we were lumping all allocations of size 0-8 into a single
slab. This is a left-over of the unnecessary writing of the size of each
slice into the byte data as a header.

Because there isn't any slab overhead in having slabs specifically set
up to allocate small slices we can now do that.

Now we have a clean and straight forward slab for each power of 2 sized
group.