Fixed: freed slots were repeatedly allocated
Because we maintain the invariant that a freed slot's meta-data always
has a non-nil 'nextFree' pointer, the last slot in the free linked-list
points to itself. We were not checking for this case, and when freeing a
slot we would simply copy the 'nextFree' from the freed slot straight to
the 'nextFree' root in the byteSlab. This meant that the last freed slot
would be allocated again and again (forever). Which is really bad.

Now we check for this case and set the 'nextFree' root to be nil when we
free the final slot.

This was uncovered while writing extra tests. Although it's good that we
found this bug, it was basically accidental and a very similar looking
set of tests would have missed this very severe bug.

I think that some kind of fuzz based testing is really needed to
generate a large corpus of more complex allocs/gets/frees than will be
provided by hand-written unit tests. But that can come later.